Refactor: 상세 페이지 컴포넌트 분리
- 정보, 탭, 탭별 컨텐츠 컴퐆너트 분리
- 탭 클릭시 로직 변경
